requirements.txt
aiogram==2.25.1
discord.py==2.2.2
google_api_python_client==2.84.0
gspread==5.7.2
oauth2client==4.1.3
pandas==1.4.2
pyTelegramBotAPI==4.10.0

telebot